backend/requirements.txt
amqp==5.3.1 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyascii==0.3.3 ; python_version >= "3.12" and python_version < "4.0"
asgiref==3.11.0 ; python_version >= "3.12" and python_version < "4.0"
beautifulsoup4==4.14.3 ; python_version >= "3.12" and python_version < "4.0"
billiard==4.2.4 ; python_version >= "3.12" and python_version < "4.0"
celery==5.6.2 ; python_version >= "3.12" and python_version < "4.0"
certifi==2026.1.4 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.4.4 ; python_version >= "3.12" and python_version < "4.0"
click-didyoumean==0.3.1 ; python_version >= "3.12" and python_version < "4.0"
click-plugins==1.1.1.2 ; python_version >= "3.12" and python_version < "4.0"
click-repl==0.3.0 ; python_version >= "3.12" and python_version < "4.0"
click==8.3.1 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and platform_system == "Windows"
defusedxml==0.7.1 ; python_version >= "3.12" and python_version < "4.0"
django-appconf==1.2.0 ; python_version >= "3.12" and python_version < "4.0"
django-autoslug==1.9.9 ; python_version >= "3.12" and python_version < "4.0"
django-cities-light==3.10.2 ; python_version >= "3.12" and python_version < "4.0"
django-countries==8.2.0 ; python_version >= "3.12" and python_version < "4.0"
django-environ==0.12.0 ; python_version >= "3.12" and python_version < "4.0"
django-filter==25.2 ; python_version >= "3.12" and python_version < "4.0"
django-modelcluster==6.4.1 ; python_version >= "3.12" and python_version < "4.0"
django-modeltranslation==0.19.19 ; python_version >= "3.12" and python_version < "4.0"
django-ninja==1.5.3 ; python_version >= "3.12" and python_version < "4.0"
django-permissionedforms==0.1 ; python_version >= "3.12" and python_version < "4.0"
django-redis==6.0.0 ; python_version >= "3.12" and python_version < "4.0"
django-select2==8.4.8 ; python_version >= "3.12" and python_version < "4.0"
django-stubs-ext==5.2.9 ; python_version >= "3.12" and python_version < "4.0"
django-taggit==6.1.0 ; python_version >= "3.12" and python_version < "4.0"
django-tasks==0.9.0 ; python_version >= "3.12" and python_version < "4.0"
django-treebeard==4.8.0 ; python_version >= "3.12" and python_version < "4.0"
django-unfold==0.78.1 ; python_version >= "3.12" and python_version < "4.0"
django==6.0.1 ; python_version >= "3.12" and python_version < "4.0"
djangorestframework==3.16.1 ; python_version >= "3.12" and python_version < "4.0"
draftjs-exporter==5.2.0 ; python_version >= "3.12" and python_version < "4.0"
et-xmlfile==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
filetype==1.2.0 ; python_version >= "3.12" and python_version < "4.0"
idna==3.11 ; python_version >= "3.12" and python_version < "4.0"
kombu==5.6.2 ; python_version >= "3.12" and python_version < "4.0"
laces==0.1.2 ; python_version >= "3.12" and python_version < "4.0"
modelsearch==1.1.1 ; python_version >= "3.12" and python_version < "4.0"
openpyxl==3.1.5 ; python_version >= "3.12" and python_version < "4.0"
packaging==26.0 ; python_version >= "3.12" and python_version < "4.0"
pillow-heif==1.2.0 ; python_version >= "3.12" and python_version < "4.0"
pillow==12.1.0 ; python_version >= "3.12" and python_version < "4.0"
progressbar2==4.5.0 ; python_version >= "3.12" and python_version < "4.0"
prompt-toolkit==3.0.52 ; python_version >= "3.12" and python_version < "4.0"
psycopg2-binary==2.9.11 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.41.5 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.12.5 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
python-utils==3.9.1 ; python_version >= "3.12" and python_version < "4.0"
pytz==2025.2 ; python_version >= "3.12" and python_version < "4.0"
redis==7.1.1 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.5 ; python_version >= "3.12" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.12" and python_version < "4.0"
soupsieve==2.8.3 ; python_version >= "3.12" and python_version < "4.0"
sqlparse==0.5.5 ; python_version >= "3.12" and python_version < "4.0"
telepath==0.3.1 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.15.0 ; python_version >= "3.12" and python_version < "4.0"
typing-inspection==0.4.2 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2025.3 ; python_version >= "3.12" and python_version < "4.0"
tzlocal==5.3.1 ; python_version >= "3.12" and python_version < "4.0"
unidecode==1.4.0 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.6.3 ; python_version >= "3.12" and python_version < "4.0"
vine==5.1.0 ; python_version >= "3.12" and python_version < "4.0"
wagtail==7.2.1 ; python_version >= "3.12" and python_version < "4.0"
wagtailmedia==0.17.2 ; python_version >= "3.12" and python_version < "4.0"
wcwidth==0.6.0 ; python_version >= "3.12" and python_version < "4.0"
whitenoise==6.11.0 ; python_version >= "3.12" and python_version < "4.0"
willow==1.12.0 ; python_version >= "3.12" and python_version < "4.0"
